+++ packages/dynamic-bonding-curve/src/helpers/buildCurve.ts
@@ -5,11 +5,10 @@ import {
     type BuildCurveParam,
     BuildCurveWithMarketCapParam,
     BuildCurveWithLiquidityWeightsParam,
-    BuildCurveWithCreatorFirstBuyParam,
     BuildCurveWithTwoSegmentsParam,
     BuildCurveWithFlatSegmentParam,
 } from '../types'
-import { FEE_DENOMINATOR, MAX_SQRT_PRICE } from '../constants'
+import { MAX_SQRT_PRICE } from '../constants'
 import {
     getSqrtPriceFromPrice,
     getMigrationBaseToken,
@@ -17,7 +16,6 @@ import {
     getFirstCurve,
     getTotalSupplyFromCurve,
     getPercentageSupplyOnMigration,
-    getMigrationQuoteThreshold,
     getSqrtPriceFromMarketCap,
     getBaseTokenForSwap,
     getSwapAmountWithBuffer,
@@ -26,9 +24,12 @@ import {
     getBaseFeeParams,
     getLockedVestingParams,
     getFlatCurve,
+    getMigrationQuoteAmountFromMigrationQuoteThreshold,
+    getMigrationQuoteAmount,
+    getMigrationQuoteThresholdFromMigrationQuoteAmount,
 } from './common'
 import { getInitialLiquidityFromDeltaBase } from '../math/curve'
-import { bpsToFeeNumerator, convertDecimalToBN } from './utils'
+import { convertDecimalToBN, fromDecimalToBN } from './utils'
 
 /**
  * Build a custom constant product curve
@@ -36,7 +37,7 @@ import { bpsToFeeNumerator, convertDecimalToBN } from './utils'
  * @returns The build custom constant product curve
  */
 export function buildCurve(buildCurveParam: BuildCurveParam): ConfigParameters {
-    const {
+    let {
         totalTokenSupply,
         percentageSupplyOnMigration,
         migrationQuoteThreshold,
@@ -54,6 +55,8 @@ export function buildCurve(buildCurveParam: BuildCurveParam): ConfigParameters {
         creatorLockedLpPercentage,
         creatorTradingFeePercentage,
         leftover,
+        tokenUpdateAuthority,
+        migrationFee,
     } = buildCurveParam
 
     const {
@@ -86,8 +89,7 @@ export function buildCurve(buildCurveParam: BuildCurveParam): ConfigParameters {
         cliffUnlockAmount,
         totalVestingDuration,
         cliffDurationFromMigrationTime,
-        tokenBaseDecimal,
-        activationType
+        tokenBaseDecimal
     )
 
     const migrationBaseSupply = new BN(totalTokenSupply)
@@ -98,12 +100,18 @@ export function buildCurve(buildCurveParam: BuildCurveParam): ConfigParameters {
         new BN(10).pow(new BN(tokenBaseDecimal))
     )
 
-    const migrationQuoteThresholdWithDecimals = new BN(
-        migrationQuoteThreshold * 10 ** tokenQuoteDecimal
+    const migrationQuoteAmount =
+        getMigrationQuoteAmountFromMigrationQuoteThreshold(
+            new Decimal(migrationQuoteThreshold),
+            migrationFee.feePercentage
+        )
+
+    const migrationPrice = new Decimal(migrationQuoteAmount.toString()).div(
+        new Decimal(migrationBaseSupply.toString())
     )
 
-    const migrationPrice = new Decimal(migrationQuoteThreshold.toString()).div(
-        new Decimal(migrationBaseSupply.toString())
+    let migrationQuoteThresholdInLamport = new BN(
+        migrationQuoteThreshold * 10 ** tokenQuoteDecimal
     )
 
     const totalLeftover = new BN(leftover).mul(
@@ -115,9 +123,12 @@ export function buildCurve(buildCurveParam: BuildCurveParam): ConfigParameters {
         tokenBaseDecimal,
         tokenQuoteDecimal
     )
+    let migrationQuoteAmountInLamport = new BN(
+        migrationQuoteThreshold * 10 ** tokenQuoteDecimal
+    )
 
     const migrationBaseAmount = getMigrationBaseToken(
-        new BN(migrationQuoteThresholdWithDecimals),
+        new BN(migrationQuoteAmountInLamport),
         migrateSqrtPrice,
         migrationOption
     )
@@ -133,16 +144,18 @@ export function buildCurve(buildCurveParam: BuildCurveParam): ConfigParameters {
         migrateSqrtPrice,
         migrationBaseAmount,
         swapAmount,
-        migrationQuoteThresholdWithDecimals
+        migrationQuoteThresholdInLamport,
+        migrationFee.feePercentage
     )
 
     const totalDynamicSupply = getTotalSupplyFromCurve(
-        migrationQuoteThresholdWithDecimals,
+        migrationQuoteThresholdInLamport,
         sqrtStartPrice,
         curve,
         lockedVesting,
         migrationOption,
-        totalLeftover
+        totalLeftover,
+        migrationFee.feePercentage
     )
 
     const remainingAmount = totalSupply.sub(totalDynamicSupply)
@@ -174,7 +187,7 @@ export function buildCurve(buildCurveParam: BuildCurveParam): ConfigParameters {
         migrationOption: migrationOption,
         tokenType: tokenType,
         tokenDecimal: tokenBaseDecimal,
-        migrationQuoteThreshold: migrationQuoteThresholdWithDecimals,
+        migrationQuoteThreshold: migrationQuoteThresholdInLamport,
         partnerLpPercentage: partnerLpPercentage,
         creatorLpPercentage: creatorLpPercentage,
         partnerLockedLpPercentage: partnerLockedLpPercentage,
@@ -187,6 +200,8 @@ export function buildCurve(buildCurveParam: BuildCurveParam): ConfigParameters {
             postMigrationTokenSupply: totalSupply,
         },
         creatorTradingFeePercentage,
+        tokenUpdateAuthority,
+        migrationFee,
         padding0: [],
         padding1: [],
         curve,
@@ -206,8 +221,8 @@ export function buildCurveWithMarketCap(
         initialMarketCap,
         migrationMarketCap,
         totalTokenSupply,
-        activationType,
         tokenBaseDecimal,
+        migrationFee,
     } = buildCurveWithMarketCapParam
 
     const {
@@ -224,8 +239,7 @@ export function buildCurveWithMarketCap(
         cliffUnlockAmount,
         totalVestingDuration,
         cliffDurationFromMigrationTime,
-        tokenBaseDecimal,
-        activationType
+        tokenBaseDecimal
     )
 
     const totalSupply = new BN(totalTokenSupply).mul(
@@ -239,10 +253,15 @@ export function buildCurveWithMarketCap(
         totalSupply
     )
 
-    const migrationQuoteThreshold = getMigrationQuoteThreshold(
+    const migrationQuoteAmount = getMigrationQuoteAmount(
         new Decimal(migrationMarketCap),
         new Decimal(percentageSupplyOnMigration)
     )
+    const migrationQuoteThreshold =
+        getMigrationQuoteThresholdFromMigrationQuoteAmount(
+            migrationQuoteAmount,
+            new Decimal(migrationFee.feePercentage)
+        ).toNumber()
 
     return buildCurve({
         ...buildCurveWithMarketCapParam,
@@ -278,6 +297,8 @@ export function buildCurveWithTwoSegments(
         activationType,
         dynamicFeeEnabled,
         migrationFeeOption,
+        migrationFee,
+        tokenUpdateAuthority,
     } = buildCurveWithTwoSegmentsParam
 
     const {
@@ -310,8 +331,7 @@ export function buildCurveWithTwoSegments(
         cliffUnlockAmount,
         totalVestingDuration,
         cliffDurationFromMigrationTime,
-        tokenBaseDecimal,
-        activationType
+        tokenBaseDecimal
     )
 
     let migrationBaseSupply = new BN(totalTokenSupply)
@@ -321,17 +341,25 @@ export function buildCurveWithTwoSegments(
     let totalSupply = new BN(totalTokenSupply).mul(
         new BN(10).pow(new BN(tokenBaseDecimal))
     )
-
-    // migrationMarketCap * migrationBaseSupply / totalTokenSupply
+    let migrationQuoteAmount = getMigrationQuoteAmount(
+        new Decimal(migrationMarketCap),
+        new Decimal(percentageSupplyOnMigration)
+    )
     let migrationQuoteThreshold =
-        (migrationMarketCap * percentageSupplyOnMigration) / 100
+        getMigrationQuoteThresholdFromMigrationQuoteAmount(
+            migrationQuoteAmount,
+            new Decimal(migrationFee.feePercentage)
+        )
 
-    let migrationQuoteThresholdWithDecimals = new BN(
-        migrationQuoteThreshold * 10 ** tokenQuoteDecimal
+    let migrationPrice = migrationQuoteAmount.div(
+        new Decimal(migrationBaseSupply.toString())
     )
 
-    let migrationPrice = new Decimal(migrationQuoteThreshold.toString()).div(
-        new Decimal(migrationBaseSupply.toString())
+    let migrationQuoteThresholdInLamport = fromDecimalToBN(
+        migrationQuoteThreshold.mul(new Decimal(10 ** tokenQuoteDecimal))
+    )
+    let migrationQuoteAmountInLamport = fromDecimalToBN(
+        migrationQuoteAmount.mul(new Decimal(10 ** tokenQuoteDecimal))
     )
 
     let migrateSqrtPrice = getSqrtPriceFromPrice(
@@ -341,7 +369,7 @@ export function buildCurveWithTwoSegments(
     )
 
     let migrationBaseAmount = getMigrationBaseToken(
-        new BN(migrationQuoteThresholdWithDecimals),
+        migrationQuoteAmountInLamport,
         migrateSqrtPrice,
         migrationOption
     )
@@ -387,13 +415,14 @@ export function buildCurveWithTwoSegments(
     let midPrices = [midSqrtPrice1, midSqrtPrice2, midSqrtPrice3]
     let sqrtStartPrice = new BN(0)
     let curve: { sqrtPrice: BN; liquidity: BN }[] = []
+
     for (let i = 0; i < midPrices.length; i++) {
         const result = getTwoCurve(
             migrateSqrtPrice,
             midPrices[i],
             initialSqrtPrice,
             swapAmount,
-            migrationQuoteThresholdWithDecimals
+            migrationQuoteThresholdInLamport
         )
         if (result.isOk) {
             curve = result.curve
@@ -403,12 +432,13 @@ export function buildCurveWithTwoSegments(
     }
 
     let totalDynamicSupply = getTotalSupplyFromCurve(
-        migrationQuoteThresholdWithDecimals,
+        migrationQuoteThresholdInLamport,
         sqrtStartPrice,
         curve,
         lockedVesting,
         migrationOption,
-        totalLeftover
+        totalLeftover,
+        migrationFee.feePercentage
     )
 
     if (totalDynamicSupply.gt(totalSupply)) {
@@ -433,7 +463,7 @@ export function buildCurveWithTwoSegments(
         migrationOption,
         tokenType,
         tokenDecimal: tokenBaseDecimal,
-        migrationQuoteThreshold: migrationQuoteThresholdWithDecimals,
+        migrationQuoteThreshold: migrationQuoteThresholdInLamport,
         partnerLpPercentage,
         creatorLpPercentage,
         partnerLockedLpPercentage,
@@ -449,6 +479,8 @@ export function buildCurveWithTwoSegments(
         padding0: [],
         padding1: [],
         curve,
+        tokenUpdateAuthority,
+        migrationFee,
     }
     return instructionParams
 }
@@ -699,7 +731,7 @@ export function buildCurveWithFlatSegment(
 export function buildCurveWithLiquidityWeights(
     buildCurveWithLiquidityWeightsParam: BuildCurveWithLiquidityWeightsParam
 ): ConfigParameters {
-    const {
+    let {
         totalTokenSupply,
         migrationOption,
         tokenBaseDecimal,
@@ -718,6 +750,8 @@ export function buildCurveWithLiquidityWeights(
         initialMarketCap,
         migrationMarketCap,
         liquidityWeights,
+        migrationFee,
+        tokenUpdateAuthority,
     } = buildCurveWithLiquidityWeightsParam
 
     const {
@@ -750,8 +784,7 @@ export function buildCurveWithLiquidityWeights(
         cliffUnlockAmount,
         totalVestingDuration,
         cliffDurationFromMigrationTime,
-        tokenBaseDecimal,
-        activationType
+        tokenBaseDecimal
     )
 
     // 1. finding Pmax and Pmin
@@ -796,18 +829,32 @@ export function buildCurveWithLiquidityWeights(
         .sub(totalVestingAmount)
         .sub(totalLeftover)
 
+    // Swap_Amount = sum(li * (1/p(i-1) - 1/pi))
+    // Quote_Amount = sum(li * (pi-p(i-1)))
+    // Quote_Amount * (1-migrationFee/100) / Base_Amount = Pmax ^ 2
+
+    // -> Base_Amount = Quote_Amount * (1-migrationFee) / Pmax ^ 2
+    // -> Swap_Amount + Base_Amount = sum(li * (1/p(i-1) - 1/pi)) + sum(li * (pi-p(i-1))) * (1-migrationFee) / Pmax ^ 2
+    // l0 * sum_factor = Swap_Amount + Base_Amount
+    // => l0 * sum_factor = sum(li * (1/p(i-1) - 1/pi)) + sum(li * (pi-p(i-1))) * (1-migrationFee/100) / Pmax ^ 2
+    // => l0 = (Swap_Amount + Base_Amount ) / sum_factor
     let sumFactor = new Decimal(0)
     let pmaxWeight = new Decimal(pMax.toString())
+    let migrationFeeFactor = new Decimal(100)
+        .sub(new Decimal(migrationFee.feePercentage))
+        .div(new Decimal(100))
     for (let i = 1; i < 17; i++) {
         let pi = new Decimal(sqrtPrices[i].toString())
         let piMinus = new Decimal(sqrtPrices[i - 1].toString())
         let k = new Decimal(liquidityWeights[i - 1])
         let w1 = pi.sub(piMinus).div(pi.mul(piMinus))
-        let w2 = pi.sub(piMinus).div(pmaxWeight.mul(pmaxWeight))
+        let w2 = pi
+            .sub(piMinus)
+            .mul(migrationFeeFactor)
+            .div(pmaxWeight.mul(pmaxWeight))
         let weight = k.mul(w1.add(w2))
         sumFactor = sumFactor.add(weight)
     }
-
     let l1 = new Decimal(totalSwapAndMigrationAmount.toString()).div(sumFactor)
 
     // construct curve
@@ -830,19 +877,26 @@ export function buildCurveWithLiquidityWeights(
     )
 
     let migrationAmount = totalSwapAndMigrationAmount.sub(swapBaseAmountBuffer)
-    // let percentage = migrationAmount.mul(new BN(100)).div(totalSupply)
 
-    // calculate migration threshold
-    let migrationQuoteThreshold = migrationAmount.mul(pMax).mul(pMax).shrn(128)
+    let migrationQuoteAmount = migrationAmount.mul(pMax).mul(pMax).shrn(128)
+    let migrationQuoteThreshold =
+        getMigrationQuoteThresholdFromMigrationQuoteAmount(
+            new Decimal(migrationQuoteAmount.toString()),
+            new Decimal(migrationFee.feePercentage)
+        )
+    let migrationQuoteThresholdInLamport = fromDecimalToBN(
+        migrationQuoteThreshold
+    )
 
     // sanity check
     let totalDynamicSupply = getTotalSupplyFromCurve(
-        migrationQuoteThreshold,
+        migrationQuoteThresholdInLamport,
         pMin,
         curve,
         lockedVesting,
         migrationOption,
-        totalLeftover
+        totalLeftover,
+        migrationFee.feePercentage
     )
 
     if (totalDynamicSupply.gt(totalSupply)) {
@@ -867,7 +921,7 @@ export function buildCurveWithLiquidityWeights(
         migrationOption: migrationOption,
         tokenType: tokenType,
         tokenDecimal: tokenBaseDecimal,
-        migrationQuoteThreshold,
+        migrationQuoteThreshold: migrationQuoteThresholdInLamport,
         partnerLpPercentage: partnerLpPercentage,
         creatorLpPercentage: creatorLpPercentage,
         partnerLockedLpPercentage: partnerLockedLpPercentage,
@@ -883,244 +937,8 @@ export function buildCurveWithLiquidityWeights(
         padding0: [],
         padding1: [],
         curve,
-    }
-    return instructionParams
-}
-
-/**
- * Build a custom curve with creator first buy (must be in collect fee mode == 0)
- * @param buildCurveWithCreatorFirstBuyParam - The parameters for the custom constant product curve with creator first buy
- * @returns The build custom constant product curve with creator first buy
- */
-export function buildCurveWithCreatorFirstBuy(
-    buildCurveWithCreatorFirstBuyParam: BuildCurveWithCreatorFirstBuyParam
-): ConfigParameters {
-    const {
-        totalTokenSupply,
-        migrationOption,
-        tokenBaseDecimal,
-        tokenQuoteDecimal,
-        dynamicFeeEnabled,
-        activationType,
-        collectFeeMode,
-        migrationFeeOption,
-        tokenType,
-        partnerLpPercentage,
-        creatorLpPercentage,
-        partnerLockedLpPercentage,
-        creatorLockedLpPercentage,
-        creatorTradingFeePercentage,
-        leftover,
-        initialMarketCap,
-        migrationMarketCap,
-        liquidityWeights,
-    } = buildCurveWithCreatorFirstBuyParam
-
-    const { quoteAmount, baseAmount } =
-        buildCurveWithCreatorFirstBuyParam.creatorFirstBuyOption
-
-    const {
-        startingFeeBps,
-        endingFeeBps,
-        numberOfPeriod,
-        feeSchedulerMode,
-        totalDuration,
-    } = buildCurveWithCreatorFirstBuyParam.feeSchedulerParam
-
-    const baseFeeParams = getBaseFeeParams(
-        startingFeeBps,
-        endingFeeBps,
-        feeSchedulerMode,
-        numberOfPeriod,
-        totalDuration
-    )
-
-    const {
-        totalLockedVestingAmount,
-        numberOfVestingPeriod,
-        cliffUnlockAmount,
-        totalVestingDuration,
-        cliffDurationFromMigrationTime,
-    } = buildCurveWithCreatorFirstBuyParam.lockedVestingParam
-
-    const lockedVesting = getLockedVestingParams(
-        totalLockedVestingAmount,
-        numberOfVestingPeriod,
-        cliffUnlockAmount,
-        totalVestingDuration,
-        cliffDurationFromMigrationTime,
-        tokenBaseDecimal,
-        activationType
-    )
-
-    // find Pmax and Pmin
-    let pMin = getSqrtPriceFromMarketCap(
-        initialMarketCap,
-        totalTokenSupply,
-        tokenBaseDecimal,
-        tokenQuoteDecimal
-    )
-    let pMax = getSqrtPriceFromMarketCap(
-        migrationMarketCap,
-        totalTokenSupply,
-        tokenBaseDecimal,
-        tokenQuoteDecimal
-    )
-
-    if (quoteAmount === 0 && baseAmount > 0) {
-        throw new Error('quoteAmount must be greater than 0')
-    }
-
-    // find p0 (initial price of curve)
-    let firstBuyQuoteAmount = new BN(
-        new Decimal(quoteAmount)
-            .mul(new Decimal(10).pow(tokenQuoteDecimal))
-            .toNumber()
-    )
-    let firstBuyBaseAmount = new BN(baseAmount).mul(
-        new BN(10).pow(new BN(tokenBaseDecimal))
-    )
-
-    const cliffFeeNumerator = bpsToFeeNumerator(startingFeeBps)
-    let quoteAmountAfterFee = firstBuyQuoteAmount
-        .mul(new BN(FEE_DENOMINATOR).sub(cliffFeeNumerator))
-        .div(new BN(FEE_DENOMINATOR))
-
-    let p0 = quoteAmountAfterFee.shln(128).div(firstBuyBaseAmount).div(pMin)
-    let l0 = quoteAmountAfterFee.shln(128).div(pMin.sub(p0))
-
-    if (pMin.lt(p0)) {
-        throw Error('first price is greater than initial market cap')
-    }
-
-    // construct first curve
-    let curve = [
-        {
-            sqrtPrice: pMin,
-            liquidity: l0,
-        },
-    ]
-
-    // find q^15 = pMax / pMin
-    let priceRatio = new Decimal(pMax.toString()).div(
-        new Decimal(pMin.toString())
-    )
-    let qDecimal = priceRatio.pow(new Decimal(1).div(new Decimal(15)))
-
-    // finding all prices
-    let sqrtPrices = []
-    let currentPrice = pMin
-    for (let i = 0; i < 16; i++) {
-        sqrtPrices.push(currentPrice)
-        currentPrice = convertDecimalToBN(
-            qDecimal.mul(new Decimal(currentPrice.toString()))
-        )
-    }
-
-    let totalSupply = new BN(totalTokenSupply).mul(
-        new BN(10).pow(new BN(tokenBaseDecimal))
-    )
-    let totalLeftover = new BN(leftover).mul(
-        new BN(10).pow(new BN(tokenBaseDecimal))
-    )
-
-    let totalVestingAmount = getTotalVestingAmount(lockedVesting)
-
-    let totalSwapAndMigrationAmount = totalSupply
-        .sub(totalVestingAmount)
-        .sub(totalLeftover)
-    let totalSwapAndMigrationAmountAfterFirstBuyAmount =
-        totalSwapAndMigrationAmount.sub(firstBuyBaseAmount)
-
-    let sumFactor = new Decimal(0)
-    let pmaxWeight = new Decimal(pMax.toString())
-    for (let i = 1; i < 16; i++) {
-        let pi = new Decimal(sqrtPrices[i].toString())
-        let piMinus = new Decimal(sqrtPrices[i - 1].toString())
-        let k = new Decimal(liquidityWeights[i - 1])
-        let w1 = pi.sub(piMinus).div(pi.mul(piMinus))
-        let w2 = pi.sub(piMinus).div(pmaxWeight.mul(pmaxWeight))
-        let weight = k.mul(w1.add(w2))
-        sumFactor = sumFactor.add(weight)
-    }
-
-    let l1 = new Decimal(
-        totalSwapAndMigrationAmountAfterFirstBuyAmount.toString()
-    ).div(sumFactor)
-
-    // construct remaining curve
-    for (let i = 0; i < 15; i++) {
-        let k = new Decimal(liquidityWeights[i])
-        let liquidity = convertDecimalToBN(l1.mul(k))
-        let sqrtPrice = i < 15 ? sqrtPrices[i + 1] : pMax
-        curve.push({
-            sqrtPrice,
-            liquidity,
-        })
-    }
-
-    // reverse to calculate swap amount and migration amount
-    let swapBaseAmount = getBaseTokenForSwap(p0, pMax, curve)
-    let swapBaseAmountBuffer = getSwapAmountWithBuffer(
-        swapBaseAmount,
-        p0,
-        curve
-    )
-
-    let migrationAmount = totalSwapAndMigrationAmount.sub(swapBaseAmountBuffer)
-    // let percentage = migrationAmount.mul(new BN(100)).div(totalSupply)
-
-    // calculate migration threshold
-    let migrationQuoteThreshold = migrationAmount.mul(pMax).mul(pMax).shrn(128)
-
-    // sanity check
-    let totalDynamicSupply = getTotalSupplyFromCurve(
-        migrationQuoteThreshold,
-        p0,
-        curve,
-        lockedVesting,
-        migrationOption,
-        totalLeftover
-    )
-
-    if (totalDynamicSupply.gt(totalSupply)) {
-        // precision loss is used for leftover
-        let leftOverDelta = totalDynamicSupply.sub(totalSupply)
-        if (!leftOverDelta.lt(totalLeftover)) {
-            throw new Error('leftOverDelta must be less than totalLeftover')
-        }
-    }
-
-    const instructionParams: ConfigParameters = {
-        poolFees: {
-            baseFee: {
-                ...baseFeeParams,
-            },
-            dynamicFee: dynamicFeeEnabled
-                ? getDynamicFeeParams(endingFeeBps)
-                : null,
-        },
-        activationType,
-        collectFeeMode,
-        migrationOption,
-        tokenType,
-        tokenDecimal: tokenBaseDecimal,
-        migrationQuoteThreshold,
-        partnerLpPercentage,
-        creatorLpPercentage,
-        partnerLockedLpPercentage,
-        creatorLockedLpPercentage,
-        sqrtStartPrice: p0,
-        lockedVesting,
-        migrationFeeOption,
-        tokenSupply: {
-            preMigrationTokenSupply: totalSupply,
-            postMigrationTokenSupply: totalSupply,
-        },
-        creatorTradingFeePercentage,
-        padding0: [],
-        padding1: [],
-        curve,
+        migrationFee,
+        tokenUpdateAuthority,
     }
     return instructionParams
 }